Improve font effect rasterization quality
Three approximations in the rasterizer showed up as artifacts rather than as
speed.

The blur was three integer box passes. Its effective width jumped whenever
the derived radius crossed an integer, so animating a blur stepped, and
large glows kept boxy corners. It is now a separable Gaussian kernel while
the radius stays small, and four fractional box passes beyond that, where
the two edge samples are weighted by the fractional part so the width varies
continuously. Both stay linear in the number of pixels; the cutoff keeps the
kernel from growing to a hundred taps on a large glow.

The signed distance field thresholded coverage at one half and then patched
the result back for pixels within one texel of the edge. That threw away the
subpixel coverage FreeType had already computed and left a discontinuity
where the patch stopped applying. The Felzenszwalb transform minimises
squared distance plus an arbitrary non negative seed, so the subpixel offset
can be seeded exactly instead: the squared distance from the pixel centre to
the edge, on both sides. The special case disappears and outline, spread,
bevel and the Distance gradient all sharpen at once, at the same cost.

Offsets rounded to whole pixels, so a shadow animated in fractional steps
jumped. fontEffectOffset now takes floats and samples bilinearly, keeping
the integer path when the offset happens to be whole.

The viewer reports microseconds per glyph next to its status line, measured
over the glyphs of the sample text, so the cost of all this is visible while
editing rather than guessed at.

+++ src/Plugins/FontEffectPlugin/FontEffectRaster.h
@@ -113,7 +113,7 @@ namespace Mengine
         void fontEffectCoverageBand( const FontEffectPlane & _sdf, float _inner, float _outer, float _sharpness, const FontEffectPlane & _out );
         void fontEffectInvert( const FontEffectPlane & _src, const FontEffectPlane & _dst );
         void fontEffectMultiply( const FontEffectPlane & _plane, const FontEffectPlane & _mask );
-        void fontEffectOffset( const FontEffectPlane & _src, int32_t _dx, int32_t _dy, const FontEffectPlane & _dst );
+        void fontEffectOffset( const FontEffectPlane & _src, float _dx, float _dy, const FontEffectPlane & _dst );
 
         void fontEffectBlurPlane( const FontEffectPlane & _plane, float _sigma, const FontEffectPlane & _tmp );
         void fontEffectBlurImage( const FontEffectImage & _image, float _sigma, const FontEffectPlane & _tmp0, const FontEffectPlane & _tmp1 );
